plugin: cleanup: remove handle_drop and round_start.
About this change: handle_drop() isn't needed anymore since we're using a
smarter method (which expects the player the dropped weaponbox to remove
it and, then, providing it with a crossbow instead).

Appart from that, round_start was broken and unused long ago, but I didn't
have the time to do a proper cleanup. I've removed it since it's got no
purpose as of now.